+++ .golangci.yml
@@ -46,6 +46,8 @@ linters-settings:
           - github.com/Serega2780/otus-golang-hw/hw12_13_14_15_calendar/internal/mapper
           - github.com/Serega2780/otus-golang-hw/hw12_13_14_15_calendar/internal/model
           - github.com/Serega2780/otus-golang-hw/hw12_13_14_15_calendar/internal/service
+          - github.com/Serega2780/otus-golang-hw/hw12_13_14_15_calendar/internal/service/memory
+          - github.com/Serega2780/otus-golang-hw/hw12_13_14_15_calendar/internal/service/sql
           - github.com/Serega2780/otus-golang-hw/hw12_13_14_15_calendar/internal/storage/sql
           - github.com/Serega2780/otus-golang-hw/hw12_13_14_15_calendar/internal/storage/repository
           - github.com/Serega2780/otus-golang-hw/hw12_13_14_15_calendar/internal/storage/memory
