+++ hw12_13_14_15_calendar/internal/server/http/middleware.go
@@ -1,4 +1,4 @@
-package internalhttp
+package http
 
 import (
 	"net/http"
