+++ hw12_13_14_15_calendar/internal/server/http/server.go
@@ -1,4 +1,4 @@
-package internalhttp
+package http
 
 import (
 	"context"
