+++ hw12_13_14_15_calendar/internal/storage/memory/storage_test.go
@@ -1,4 +1,4 @@
-package memorystorage
+package memory
 
 import (
 	"context"
